+++ tests/test_events_api.py
@@ -1,5 +1,5 @@
 from os import environ
-from typing import Optional
+from typing import Callable
 from unittest.mock import MagicMock, patch
 
 from lbz.aws_boto3 import Boto3Client
@@ -10,8 +10,8 @@ class MyTestEvent(BaseEvent):
     type = "MY_TEST_EVENT"
 
 
-class TestBaseNewEvent:
-    def test_base_new_event(self) -> None:
+class TestBaseEvent:
+    def test_base_event_creation_and_structure(self) -> None:
         new_event = MyTestEvent({"x": 1})
 
         assert hasattr(new_event, "raw_data")
@@ -22,17 +22,14 @@ class TestBaseNewEvent:
 
 
 class TestEventApi:
-    event_api: Optional[EventAPI] = None
-
     def setup_method(self) -> None:
         # pylint: disable= attribute-defined-outside-init
         with patch.dict(environ, {"AWS_LAMBDA_FUNCTION_NAME": "million-dollar-lambda"}):
             self.event_api = EventAPI()
 
-    def teardown_method(self, _test_method) -> None:
+    def teardown_method(self, _test_method: Callable) -> None:
         # pylint: disable= attribute-defined-outside-init
-        self.event_api._del()  # pylint: disable=protected-access
-        self.event_api = None
+        self.event_api._del()  # type: ignore # pylint: disable=protected-access
 
     def test_init_with_defaults(self) -> None:
         # pylint: disable=protected-access
@@ -41,6 +38,23 @@ class TestEventApi:
         assert self.event_api._source == "million-dollar-lambda"
         assert self.event_api._bus_name == "million-dollar-lambda-event-bus"
         assert self.event_api._resources == []
+
+    def test___repr__(self):
+        repr = str(self.event_api)
+        assert repr == "<EventAPI bus: million-dollar-lambda-event-bus Events: pending=0 sent=0 failed=0>"
+        self.event_api.register(MyTestEvent({"x": 1}))
+        self.event_api.register(MyTestEvent({"x": 1}))
+        repr = str(self.event_api)
+        assert repr == "<EventAPI bus: million-dollar-lambda-event-bus Events: pending=2 sent=0 failed=0>"
+        self.event_api._mark_sent()  # pylint: disable=protected-access
+        repr = str(self.event_api)
+        assert repr == "<EventAPI bus: million-dollar-lambda-event-bus Events: pending=0 sent=2 failed=0>"
+        self.event_api.register(MyTestEvent({"x": 1}))
+        self.event_api.register(MyTestEvent({"x": 1}))
+        self.event_api._mark_failed()  # pylint: disable=protected-access
+        repr = str(self.event_api)
+        assert repr == "<EventAPI bus: million-dollar-lambda-event-bus Events: pending=0 sent=2 failed=2>"
+
 
     def test_set_source(self) -> None:
         # pylint: disable=protected-access
@@ -53,11 +67,10 @@ class TestEventApi:
         assert self.event_api._resources == ["x", "Y"]
 
     def test_register(self) -> None:
-        # pylint: disable=protected-access
         event = MyTestEvent({"x": 1})
         self.event_api.register(event)
 
-        assert self.event_api._pending_events == [event]
+        assert self.event_api.get_all_pending_events() == [event]
 
     def test_get_all_registered_events(self) -> None:
         event_1 = MyTestEvent({"x": 1})
@@ -67,8 +80,7 @@ class TestEventApi:
         assert self.event_api.get_all_pending_events() == [event_1, event_2]
 
     @patch.object(Boto3Client, "eventbridge")
-    @patch.object(EventAPI, "_mark_sent")
-    def test_send(self, mock__mark_sent: MagicMock, mock_send: MagicMock) -> None:
+    def test_send(self, mock_send: MagicMock) -> None:
         event = MyTestEvent({"x": 1})
         self.event_api.register(event)
 
@@ -85,25 +97,43 @@ class TestEventApi:
                 }
             ]
         )
-        mock__mark_sent.assert_called_once()
+        assert self.event_api.get_all_sent_events() == [event]
 
-    def test_clear(self):
-        # pylint: disable=protected-access
+    def test__mark_sent_moves_events_from_pending_to_sent(self) -> None:
         event = MyTestEvent({"x": 1})
         self.event_api.register(event)
 
-        self.event_api._mark_sent()
+        self.event_api._mark_sent()  # pylint: disable=protected-access
 
-        assert self.event_api._pending_events == []
-        assert self.event_api._sent_events == [event]
+        assert self.event_api.get_all_pending_events() == []
+        assert self.event_api.get_all_sent_events() == [event]
+        assert self.event_api.get_all_failed_events() == []
 
-    def test_get_all_sent_events(self):
+    def test__mark_failed_moves_events_from_pending_to_failed(self) -> None:
+        event = MyTestEvent({"x": 1})
+        self.event_api.register(event)
+
+        self.event_api._mark_failed()  # pylint: disable=protected-access
+
+        assert self.event_api.get_all_pending_events() == []
+        assert self.event_api.get_all_sent_events() == []
+        assert self.event_api.get_all_failed_events() == [event]
+
+    def test_get_all_sent_events(self) -> None:
         event = MyTestEvent({"x": 1})
         self.event_api.register(event)
 
         self.event_api._mark_sent()  # pylint: disable=protected-access
 
         assert self.event_api.get_all_sent_events() == [event]
+
+    def test_get_all_failed_events(self) -> None:
+        event = MyTestEvent({"x": 1})
+        self.event_api.register(event)
+
+        self.event_api._mark_failed()  # pylint: disable=protected-access
+
+        assert self.event_api.get_all_failed_events() == [event]
 
     @patch.object(Boto3Client, "eventbridge")
     def test_send_no_events(self, mock_send: MagicMock) -> None:
@@ -124,14 +154,12 @@ class TestEventApi:
             "Source": "million-dollar-lambda",
         }
 
-
-class TestEventApiSingleton:
     def test_singleton_pattern_working_correctly_for_event_api(self) -> None:
         event_api_1 = EventAPI()
         event_api_2 = EventAPI()
         event_api_2.set_source("XXX")
         event_api_2.set_resources(["a", "b"])
         event_api_3 = EventAPI()
+        assert event_api_1 is event_api_2 is event_api_3 is self.event_api
         assert event_api_3._source == "XXX"  # pylint: disable=protected-access
-        assert event_api_3._resources == ["a", "b"]  # pylint: disable=protected-access
-        assert event_api_1 == event_api_2 == event_api_3
+        assert self.event_api._resources == ["a", "b"]  # pylint: disable=protected-access